Refactoring of label splitting; new unit test still failing.
